add UserPreferences.kt
This class act as a main class for implementing data store preferences. We have a function `saveAuth` which stores the email once the user logs in. Initially there is nothing in the datastore so it's value is null.

The other variable is `authToken` which is used to store the values when the user Logs in for the first time.